smart-tactic-prod/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Database dependencies
google-cloud-firestore==2.11.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7

# Google Cloud dependencies
google-cloud-logging==3.5.0
google-cloud-secret-manager==2.16.4
google-generativeai==0.3.2

# LLM and AI dependencies
langfuse==2.0.0

# Configuration and environment
python-dotenv==1.0.0

# Validation and data processing
email-validator==2.0.0

# Monitoring and observability
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-exporter-google-cloud-trace==1.20.0

# Utilities
requests==2.31.0
urllib3==2.0.4

# Development and testing (optional)
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
